+++ spring-cloud-tencent-commons/src/test/java/com/tencent/cloud/common/metadata/config/MetadataAutoConfigurationTest.java
@@ -13,14 +13,13 @@
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
+ *
  */
 
 package com.tencent.cloud.common.metadata.config;
 
 import com.tencent.cloud.common.metadata.filter.gateway.MetadataFirstScgFilter;
 import com.tencent.cloud.common.metadata.filter.gateway.MetadataFirstZuulFilter;
-import com.tencent.cloud.common.metadata.filter.web.MetadataReactiveFilter;
-import com.tencent.cloud.common.metadata.filter.web.MetadataServletFilter;
 import com.tencent.cloud.common.metadata.interceptor.feign.MetadataFirstFeignInterceptor;
 import org.assertj.core.api.Assertions;
 import org.junit.Test;
@@ -53,14 +52,6 @@ public class MetadataAutoConfigurationTest {
 				.run(context -> {
 					Assertions.assertThat(context)
 							.hasSingleBean(MetadataLocalProperties.class);
-					Assertions.assertThat(context).doesNotHaveBean(
-							MetadataAutoConfiguration.MetadataServletFilterConfig.class);
-					Assertions.assertThat(context)
-							.doesNotHaveBean(MetadataServletFilter.class);
-					Assertions.assertThat(context).doesNotHaveBean(
-							MetadataAutoConfiguration.MetadataReactiveFilterConfig.class);
-					Assertions.assertThat(context)
-							.doesNotHaveBean(MetadataReactiveFilter.class);
 					Assertions.assertThat(context).hasSingleBean(
 							MetadataAutoConfiguration.MetadataFeignInterceptorConfig.class);
 					Assertions.assertThat(context)
@@ -87,14 +78,6 @@ public class MetadataAutoConfigurationTest {
 					Assertions.assertThat(context)
 							.hasSingleBean(MetadataLocalProperties.class);
 					Assertions.assertThat(context).hasSingleBean(
-							MetadataAutoConfiguration.MetadataServletFilterConfig.class);
-					Assertions.assertThat(context)
-							.hasSingleBean(MetadataServletFilter.class);
-					Assertions.assertThat(context).doesNotHaveBean(
-							MetadataAutoConfiguration.MetadataReactiveFilterConfig.class);
-					Assertions.assertThat(context)
-							.doesNotHaveBean(MetadataReactiveFilter.class);
-					Assertions.assertThat(context).hasSingleBean(
 							MetadataAutoConfiguration.MetadataFeignInterceptorConfig.class);
 					Assertions.assertThat(context)
 							.hasSingleBean(MetadataFirstFeignInterceptor.class);
@@ -119,14 +102,6 @@ public class MetadataAutoConfigurationTest {
 				.run(context -> {
 					Assertions.assertThat(context)
 							.hasSingleBean(MetadataLocalProperties.class);
-					Assertions.assertThat(context).doesNotHaveBean(
-							MetadataAutoConfiguration.MetadataServletFilterConfig.class);
-					Assertions.assertThat(context)
-							.doesNotHaveBean(MetadataServletFilter.class);
-					Assertions.assertThat(context).hasSingleBean(
-							MetadataAutoConfiguration.MetadataReactiveFilterConfig.class);
-					Assertions.assertThat(context)
-							.hasSingleBean(MetadataReactiveFilter.class);
 					Assertions.assertThat(context).hasSingleBean(
 							MetadataAutoConfiguration.MetadataFeignInterceptorConfig.class);
 					Assertions.assertThat(context)